# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/paths": {
      "get": {
        "description": "Retrieves a paginated list of the signed-in author’s Paths.",
        "operationId": "PathController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of paths retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "slug": {
                        "type": "string",
                        "example": "my-awesome-path"
                      },
                      "title": {
                        "type": "string",
                        "example": "My Awesome Path"
                      },
                      "description": {
                        "type": "string",
                        "example": "A journey through code"
                      },
                      "type": {
                        "type": "string",
                        "example": "living"
                      },
                      "visibility": {
                        "type": "string",
                        "example": "public"
                      },
                      "kind": {
                        "type": "string",
                        "example": "guide"
                      },
                      "entry": {
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                      },
                      "authorId": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "homeId": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "meta": {
                        "type": "object",
                        "example": {}
                      },
                      "created": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "updated": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Get all paths",
        "tags": [
          "Paths"
        ]
      },
      "post": {
        "description": "Creates an empty owned Path with a generated ID and no entry marker. Add markers before selecting an entry.",
        "operationId": "PathController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePathDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The path has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "slug": {
                      "type": "string",
                      "example": "my-awesome-path"
                    },
                    "title": {
                      "type": "string",
                      "example": "My Awesome Path"
                    },
                    "description": {
                      "type": "string",
                      "example": "A journey through code"
                    },
                    "type": {
                      "type": "string",
                      "example": "living"
                    },
                    "visibility": {
                      "type": "string",
                      "example": "unlisted"
                    },
                    "kind": {
                      "type": "string",
                      "example": "guide"
                    },
                    "entry": {
                      "type": "string",
                      "format": "uuid",
                      "nullable": true
                    },
                    "authorId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "homeId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "created": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updated": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient permissions"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new path",
        "tags": [
          "Paths"
        ]
      }
    },
    "/paths/{identifier}": {
      "get": {
        "description": "Retrieves an owned Path by UUID or an author-scoped slug.",
        "operationId": "PathController_getByIdentifier",
        "parameters": [
          {
            "name": "identifier",
            "required": true,
            "in": "path",
            "description": "The UUID or slug of the path",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The path data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "slug": {
                      "type": "string",
                      "example": "my-awesome-path"
                    },
                    "title": {
                      "type": "string",
                      "example": "My Awesome Path"
                    },
                    "description": {
                      "type": "string",
                      "example": "A journey through code"
                    },
                    "type": {
                      "type": "string",
                      "example": "living"
                    },
                    "visibility": {
                      "type": "string",
                      "example": "public"
                    },
                    "kind": {
                      "type": "string",
                      "example": "guide"
                    },
                    "entry": {
                      "type": "string",
                      "format": "uuid",
                      "nullable": true
                    },
                    "authorId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "homeId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "created": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updated": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Path not found"
          }
        },
        "summary": "Get a path by identifier",
        "tags": [
          "Paths"
        ]
      }
    },
    "/paths/{id}": {
      "patch": {
        "description": "Updates an existing path with the provided data. Requires authentication and ownership.",
        "operationId": "PathController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The UUID of the path to update",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePathDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The path has been successfully updated.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "slug": {
                      "type": "string",
                      "example": "my-updated-path"
                    },
                    "title": {
                      "type": "string",
                      "example": "My Updated Path"
                    },
                    "description": {
                      "type": "string",
                      "example": "Updated description"
                    },
                    "type": {
                      "type": "string",
                      "example": "frozen"
                    },
                    "visibility": {
                      "type": "string",
                      "example": "public"
                    },
                    "kind": {
                      "type": "string",
                      "example": "wander"
                    },
                    "entry": {
                      "type": "string",
                      "format": "uuid",
                      "nullable": true
                    },
                    "authorId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "homeId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "created": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updated": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "Path not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a path",
        "tags": [
          "Paths"
        ]
      },
      "delete": {
        "description": "Deletes an existing path. Requires authentication and ownership.",
        "operationId": "PathController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The UUID of the path to delete",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The path has been successfully deleted."
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "Path not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a path",
        "tags": [
          "Paths"
        ]
      }
    },
    "/paths/{id}/markers": {
      "get": {
        "description": "Retrieves all markers (crux references) for a specific path in order.",
        "operationId": "PathController_getMarkers",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The UUID of the path",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Markers retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "pathId": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "cruxId": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "order": {
                        "type": "number",
                        "example": 0
                      },
                      "note": {
                        "type": "string",
                        "example": "Start here"
                      },
                      "authorId": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "created": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "updated": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Path not found"
          }
        },
        "summary": "Get all markers for a path",
        "tags": [
          "Paths"
        ]
      },
      "put": {
        "description": "Replaces all markers for a path with the provided list. Each marker references a crux by ID.",
        "operationId": "PathController_syncMarkers",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The UUID of the path",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncMarkersDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Markers synced successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "pathId": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "cruxId": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "order": {
                        "type": "number",
                        "example": 0
                      },
                      "note": {
                        "type": "string",
                        "example": "Start here"
                      },
                      "authorId": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "created": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "updated": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid marker data"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "No permission to manage markers for this path"
          },
          "404": {
            "description": "Path or referenced crux not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Sync markers for a path",
        "tags": [
          "Paths"
        ]
      }
    },
    "/paths/{id}/tags": {
      "get": {
        "description": "Retrieves all tags associated with a specific path.",
        "operationId": "PathController_getTags",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The UUID of the path",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          },
          {
            "name": "filter",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tags retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "label": {
                        "type": "string",
                        "example": "tutorial"
                      },
                      "resourceType": {
                        "type": "string",
                        "example": "path"
                      },
                      "resourceId": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "authorId": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "homeId": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "created": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "updated": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Path not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all tags for a path",
        "tags": [
          "Paths"
        ]
      },
      "put": {
        "description": "Replaces all tags for a path with the provided list. Tags must be in kebab-case format.",
        "operationId": "PathController_syncTags",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The UUID of the path",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "labels": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": [
                      "tutorial",
                      "learning-path",
                      "beginner"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tags synced successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "label": {
                        "type": "string",
                        "example": "tutorial"
                      },
                      "resourceType": {
                        "type": "string",
                        "example": "path"
                      },
                      "resourceId": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "authorId": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "homeId": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "created": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "updated": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid tag label format (must be kebab-case)"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "No permission to manage tags for this path"
          },
          "404": {
            "description": "Path not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Sync tags for a path",
        "tags": [
          "Paths"
        ]
      }
    },
...
  },
  "info": {
    "title": "Crux Garden API",
    "description": "",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreatePathDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "description": "Unique slug identifier for the path",
            "example": "my-awesome-path",
            "minLength": 1
          },
          "title": {
            "type": "string",
            "description": "Optional title for the path",
            "example": "My Awesome Path"
          },
          "description": {
            "type": "string",
            "description": "Optional description for the path",
            "example": "This is a description of my path"
          },
          "type": {
            "type": "string",
            "description": "Type of the path",
            "enum": [
              "living",
              "frozen"
            ],
            "example": "living",
            "default": "living"
          },
          "visibility": {
            "type": "string",
            "description": "Visibility setting for the path",
            "enum": [
              "public",
              "private",
              "unlisted"
            ],
            "example": "unlisted",
            "default": "unlisted"
          },
          "kind": {
            "type": "string",
            "description": "Kind of path (guide or wander)",
            "enum": [
              "guide",
              "wander"
            ],
            "example": "wander"
          }
        },
        "required": [
          "slug",
          "kind"
        ]
      },
      "UpdatePathDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Updated title for the path",
            "example": "My Updated Path Title"
          },
          "description": {
            "type": "string",
            "description": "Updated description for the path",
            "example": "Updated description for my path"
          },
          "type": {
            "type": "string",
            "description": "Updated type of the path",
            "enum": [
              "living",
              "frozen"
            ],
            "example": "frozen"
          },
          "visibility": {
            "type": "string",
            "description": "Updated visibility setting for the path",
            "enum": [
              "public",
              "private",
              "unlisted"
            ],
            "example": "public"
          },
          "kind": {
            "type": "string",
            "description": "Updated kind of path (guide or wander)",
            "enum": [
              "guide",
              "wander"
            ],
            "example": "guide"
          },
          "entry": {
            "type": "string",
            "description": "Updated entry marker ID for the path",
            "example": "550e8400-e29b-41d4-a716-446655440001"
          }
        }
      },
      "MarkerInput": {
        "type": "object",
        "properties": {
          "cruxId": {
            "type": "string",
            "description": "Crux ID for the marker",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "order": {
            "type": "number",
            "description": "Order/position of the marker in the path",
            "example": 0
          },
          "note": {
            "type": "string",
            "description": "Optional note for the marker",
            "example": "This is an important marker"
          }
        },
        "required": [
          "cruxId",
          "order"
        ]
      },
      "SyncMarkersDto": {
        "type": "object",
        "properties": {
          "markers": {
            "description": "Array of markers to sync for the path",
            "example": [
              {
                "cruxId": "550e8400-e29b-41d4-a716-446655440000",
                "order": 0,
                "note": "First marker"
              },
              {
                "cruxId": "660e8400-e29b-41d4-a716-446655440000",
                "order": 1
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MarkerInput"
            }
          }
        },
        "required": [
          "markers"
        ]
      },
...
    }
  }
}